Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> red bots

I looked at the images and the version in the codebase is definitely wrong. It doesn't even have the start and stop patterns of PDF417. The new one does, but it's still wrong as I tried a number of decoders and neither could scan it.

Updated the hash, but perhaps we should just disable this test until it's completely fixed. Also filed pdfium:1135 to look at that.

Patch-set: 4
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>