Use MaybeOwned<> in CFGAS_GEFont.
In turn, this requires cleaning up the stock fonts after the
MaybeOwned<>, which can't happen as explicit logic in the document
destructor, because that runs too early. Instead, create a new
cleaner class to do this work for us as part of its destructor,
and arrange for that to happen after the MaybeOwned<> is gone.

Use Optional<> while we're at it in place of a bool/int pair.

Change-Id: I039aaebefd0b96c735725c5db25d0919763ca2e2