Handle non-printing characters at beginning of extraction region
Currently if a text extraction region begins on a non-printing
character then "" will be returned. This is the incorrect behaviour,
instead the call should scan ahead until a printing character is
found and start extracting from there. Also proactively adds a
similar check and scan for the end of the extraction region.

BUG=pdfium:1139

Change-Id: Ia2001ac89740f3d31d2bb69e8000773f8b01091b