Add unit test for AutoRestorer<> on smart pointer types.
Test UnownedPtr<> and RetainPtr<>. We cannot operate against
std::unique_ptr<> because we need a copyable type. Add test for
self-reassignment on going out of scope. Move one test helper
class to testing/ so it can be shared among unit tests, and
allow it to recognize if it ever gets "destroyed".

Change-Id: I13056094c70079f7283cbc7600948f81a64874b4