Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > With only a top and left value for margin values, what happens to the bottom and right margins? Do the two sides have to be symmetrical? If so, why does it have to be that way?
> > 
> > When the PDFium API was reviewed, we have decided to center the pages in their allocated area.  It doesn't have to be symmetrical, but it looks good, if it is.
> 
> If we later on decide to support custom margin, then it makes sense to input all the margins, margin_left, margin_top, margin_bottom, and margin_right.

On the PDFium side, we try not to cater to only Chromium's use cases. e.g. the PDFium API supports XxY up, not N-up for a fixed set of N-values.

If we are going to add margin controls and break the public API, can we just do that once instead of twice? If the API passes in "const FS_RECTF* margins" then all sides are adjustable.

Patch-set: 5
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>