Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > (1 comment)
> > > 
> > > > Patch Set 5:
> > > > 
> > > > > Patch Set 5:
> > > > > 
> > > > > > Patch Set 5:
> > > > > > 
> > > > > > > Patch Set 5:
> > > > > > > 
> > > > > > > With only a top and left value for margin values, what happens to the bottom and right margins? Do the two sides have to be symmetrical? If so, why does it have to be that way?
> > > > > > 
> > > > > > When the PDFium API was reviewed, we have decided to center the pages in their allocated area.  It doesn't have to be symmetrical, but it looks good, if it is.
> > > > > 
> > > > > If we later on decide to support custom margin, then it makes sense to input all the margins, margin_left, margin_top, margin_bottom, and margin_right.
> > > > 
> > > > On the PDFium side, we try not to cater to only Chromium's use cases. e.g. the PDFium API supports XxY up, not N-up for a fixed set of N-values.
> > > > 
> > > > If we are going to add margin controls and break the public API, can we just do that once instead of twice? If the API passes in "const FS_RECTF* margins" then all sides are adjustable.
> > > 
> > > The problem with using FS_RECTF struct is that, we don't know the width and height of subpages yet.  We only know the left, right, top and bottom margin.  So is't not really a FS_RECTF.
> > 
> > Just to be clear, are the margin values for the entire sheet, or for each individual subpage?
> 
> Right now, I set to be each individual subpage.  Otherwise, when pages_per_sheet >=9, the margins maybe smaller than page_size - printable_area.

Thank you for clarifying. It was not clear from the CL description or the comments in the code that this is what the implementation does.

So if the margins are applied to subpages, then it has to do even more validation to make sure there room leftover for the subpages. e.g. for an |output_width| of 1000, if |num_pages_on_x_axis| is 4, then a |margin_left| value of 125 of greater will leave no room for the subpages.

Taking a step back, is this really what you want? e.g. with 4-up on US Letter paper and a 1-inch margin, that would mean 1 inch margin on the left, 1 inch on the right, and a 2 inch gap in the middle. So only 4.5 inches left for page content.

Patch-set: 5
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>