PrintPreview: Add margin-top and margin-left as in-parameters.
Added two in-parameters margin_top and margin_left to PDFium API
FPDF_ImportNPagesToOne.  Margin_top and margin_left will be used
to computer the scale factor of a page when it needs to be scaled
and fit into it's allocated area on the output page.

Set margin_top and margin_left to zero, if margins are not needed
in some cases.

Bug:775999
Change-Id: I9c1b19ed45e6010907606fd3a24c061554d729a1