Set correct stream index when parsing is done in several steps.
When parsing happens in several steps, the position is reset to 0 and
the start pointer is advanced. This makes the calculation for which
stream a page object that was just created belongs to incorrect.

Passing in the base pointer separately the start offset allows the
correct position to be used and the correct stream to be identified.

Change-Id: Ic0d5f59f437609158aa97b3c8a18dbd48cd3b0d4