Set correct stream index when parsing is done in several steps.
When parsing happens in several steps (in pages with > 100 page
objects), the position is reset to 0 and the start pointer is
advanced. This breaks the calculation of which stream an object
belongs to.

Passing in the base pointer separately from the start offset allows
the correct position to be calculated and the correct stream to be
identified.

Change-Id: Ic0d5f59f437609158aa97b3c8a18dbd48cd3b0d4