Optimize CPDF_CMapParser
In this parser, a vector of CodeRange values are built up during
operation that need to be sync'd with the CPDF_CMap that is being
initialized. The existing implementation the vector being built as a
member var for the parser, and copying the values over to the cmap
whenever there is a change. This copy, when profiling the code, is the
dominate factor on the run time.

The code has been rewritten to have the parser reference/modify the
instance of the vector in the cmap and removes the parser's copy. This
removes all of the copies and significatnly speeds things up.

BUG=chromium:881678

Change-Id: Ib8e75962507ca3d3b1ed066fd1faa4fbb7141122