Introduce FX_Number class as a replacement for FX_atonum().
The issue with FX_atonum() is that it doesn't return any information
about whether it range-checked its integer values as signed or
unsigned types, even though it knows this as part of its processing.

Rather than adding another out parameter to that function, create
a class to hold all this information together.

This is the first place things go astray in the referenced bug, in
that a large positive value is cast to float as a negavtive value.

Bug: 882959
Change-Id: I0977ec8fccf85e2632a962507bdd30a1cbe6d33c