Change signature of FPDFPageObjMark_Get(Name|ParamKey).
These methods used to return the size of the buffer to contain the
value to be returned. Now they will return an FPDF_BOOL to make it
possible to differentiate between a success with empty return and a
failure. The size will be returned through an out params.

This CL also adds more focused testing for these API methods and
similar ones.

Change-Id: I6f9837f99d955aaba2c49a259ed7805a286e091d