Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > With some other APIs with UTF16, a return value of 0 means failure, and a return value of 2 means empty string. i.e. L"" = {0x00, 0x00}. How do you feel about that vs returning a bool?
> > 
> > What Lei describes seems awkward and counter-intuitive to me, but it may be a cultural norm for pdfium; I don't know. A boolean return value is unambiguously out of band from the meaning of the length variable, which seems better to me as a design.
> 
> It is possible that the existing code, without this CL, can already let the caller distinguish failure from empty string. Really, we can go either way on this. Here's my reasons for advocating for the existing pattern:
> 
> - Consistency: The behavior I described feels a bit quirky, but other PDFium APIs already do this. Once a developer understand this, they don't have to deal with some APIs that do it one way vs the other.

Regarding consistency, on one side we want to be consistent with the other APIs that return "unsigned long". On the other side, we want to be consistent with the FPDFPageObjMark_GetParamStringValue and FPDFPageObjMark_GetParamBlobValue calls, which are from the same group (mark APIs), and return bools. The reason they do so is to be consistent with the FPDFPageObjMark_GetParamIntValue call. So there's a chain of consistency which must break somewhere.

I believe returning bool is better as an API - it's more intuitive for a caller to check "if (success)" than "if (len > 0)" to know if the operation was successful or not. Therefore, I think the change in this CL is positive.

> - Laziness: We don't have to change anything if we stick with the existing APIs. :)

Totally agree for the existing consolidated APIs, but for these we're at the point that we can still change them with relative ease.

> - With only an unsigned int return value, there's no possiblity of contraditions like true return value with *out_buflen set to 0.

It's a downside, but minor IMO. C APIs that return "success or not" and return the actual value through an out param are common.

Patch-set: 4
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0