Change signature of FPDFPageObjMark_Get(Name|ParamKey).
These methods used to return the size of the buffer to contain the
value to be returned. Now they will return an FPDF_BOOL to make it
consistent with the other mark APIs and more intuitive to
differentiate a success from a failure. The size will be returned
through an out param.

This CL also adds more focused testing for these API methods and
similar ones.

Change-Id: I6f9837f99d955aaba2c49a259ed7805a286e091d