Add XFA GIF pixel test
This adds a test case where a simple red square GIF image is embedded
into an XFA document. This exercises that <image> tags work correctly
and that end to end rendering of GIFs is working.

Displaying GIFs is currently broken, https://crbug.com/pdfium1167, so
this test is currently suppressed.

BUG=pdfium:904

Change-Id: I400a2a287c24726bc811dacfabe30ddc77f38b9c