Add XFA BMP pixel test
This adds a test case where a simple red square BMP image is embedded
into an XFA document. This exercises that <image> tags work correctly
and that end to end rendering of BMPs is working.

Displaying BMPs is currently broken, https://crbug.com/pdfium/1168, so
this test is currently suppressed.

BUG=pdfium:1163

Change-Id: I3ef006a7aa16717ad9cdbca1d423613f17246a5e