Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> RS lgtm, I'm in an office offsite. Is the bug fixed or why is there no expectation failing?

The bug is not fixed. The to-be-checked-in "hello world" PNG is the bad result. The correct result is a blank image.

With our testing infrastructure, unexpected success is hard to notice. Whereas rendering differences are. So I would prefer to just not suppress. Does this make sense, or am I doing it wrong?

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0