Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > RS lgtm, I'm in an office offsite. Is the bug fixed or why is there no expectation failing?
> 
> The bug is not fixed. The to-be-checked-in "hello world" PNG is the bad result. The correct result is a blank image.
> 
> With our testing infrastructure, unexpected success is hard to notice. Whereas rendering differences are. So I would prefer to just not suppress. Does this make sense, or am I doing it wrong?

Hmm I think it would be more confusing for someone who accidentally fixes the bug to find a 'regression' here, so unless you plan to work on the bug in the short term then I'd prefer we check in the correct result and suppress it.

Could you indicate in the CL description why the correct rendering result is a blank page? It's not obvious and could help whoever attempts to fix the bug.

Patch-set: 1
Reviewer: Gerrit User 6295 <6295@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Code-Review=0