Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > > Patch Set 7:
> > > 
> > > > Patch Set 5:
> > > > 
> > > > Is https://crbug.com/877165 the underlying problem that we really need to solve?
> > > 
> > > No idea.
> > 
> > On https://crbug.com/877165, it mentions concurrency, but really it is about re-entrancy.
> > 
> > We also have https://crbug.com/884122 from the same bug reporter.
> 
> Both those still repro following this change.

Right. That's expected. What I'm trying to say is, if we do not fix the reentrancy issue in https://crbug.com/877165, then lots of code have to be made reentrant. We'll be playing whack-a-mole for a long time.

Patch-set: 7
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>