Remove SetSaveCalled API
This was meant to be part of a mechanism to alert users that entered
data into a PDF form was not going to actually be saved how they would
expect. The UI for this is blocked on a bug in V8, and is now being
superseded by work to correctly implement saving.

BUG=pdfium:953

Change-Id: Id9c85c109a3f6a6b4ee69d35f366006be4dc9c32