Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > I got an [Abort Retry Ignore] dialog.
> > 
> > assert(sp->cinfo.comm.is_decompressor) failed on third_party/libtiff/tif_jpeg.c:1030.
> 
> And that's because TIFFjpeg_create_decompress() failed...

I am a little confused on why there is a dialogue appearing. I thought failing an assert should terminate the process. I assume this is a Windows-ism. It is a little annoying that the libtiff code cannot decide between returning error codes and failing asserts for error handling/reporting.

I think the only reason that create_decompress can fail is if the call into the JPEG library is failing. I am unsure why it would only be failing on these specific Win configs, since the jpeg library is listed in the DEPS for libtiff. I suspecting that this might be  a clang vs msvc issue, but I think there is both clang and msvc builds that are failing.

Any thoughts on what the difference could be between the failing and passing bots?

Patch-set: 3
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0