Remove PrintParams object.
The only way to get one is via doc.getPrintParams(), and it
always returns in error. Even if one were returend to JS, there
aren't any methods/callbacks to change its properties. Even if
there were methods to change its properties, the information it
stores doesn't align with what is in the JS spec for this object.
Even if it were aligned with the JS spec, the way we check for its
presence as the ninth parameter in CJS_Document::print() doesn't
match the way its use is documented in the spec as the sole
first parameter.

Instead implement the older, simpler behaviour as spec'd prior to
6.0, and toss this enhancement altogether.

Move routine to check the result of ExpandKeywordParams to
js_define.h so other files can use it. Update comment in that
file to remove reference to deprecated enum.

Expand test for API and introduce callback to log parameters.

Change-Id: I047e2be0d2afbad91d6b58c6c74bbea083fed330