Test color.convert() and equal() from JS (and fix comparison logic).
Currently, color.equal(a, b) may not give the same result as
color.equal(b, a) since arg1 is converted to be the type of
arg2, and some of these conversions lose information. Instead
promote to the type with the most components in the hope of
preserving the most information.

Change-Id: I1d93fa29db4fb65e0f7c07c3ba7d9ca87ebf7bc9