Rename CPDF_Object::Type::BOOLEAN to BOOLEANTY
Well, this is silly, but BOOLEAN shadows a typedef in winnt.h, and new Clang
warnings warn about this as part of -Wshadow (see bug). I tried making
CPDF_Object::Type an enum class, but some code wants to use it as an integer,
so instead let's rename BOOLEAN a little.

Bug: chromium:895475
Change-Id: I278ce9dbbd88f802fe5bc77fca28b082aa5e684d