Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> yikes.  sure abort() isn't good enough vs. need to update alla time?

It looks like abort() is not good enough for Chromium. I don't think we will need to sync all the time. Only when there are complaints that the implementation is problematic.

Contrary to the deleted TODO, this does not pull in lots of base/ dependencies, so it is somewhat do able.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0