Be more particular about FX objects constructed from JS
This is back-filling some more error cases from the work from
a few weeks ago.

Replaces a lambda with a static CallHandler since the verbosity was
increasing. This is what gets invoked if you try to make a new
FXJS object from the javascript side, rather than the C++ side.

Making such an object is a little tricky, since we don't give
these functions names V8, but they can be obtained via constructor
property from an instance of the object.


Change-Id: Ibca686e75338ac54d08a114f36f930cd424a1eb5