Make CPWL_Wnd own its pAttachedData.
This requires moving it out of CreateParams, since that is
a copyable struct, and implies that currently there is some
questionable sharing going on. To resolve this, introduce a
Clone() method so that each window gets its own copy.

Make GetAttachedData() return a const pointer, so that callers
can't free it behind our back.

Change-Id: Iadc97688b4692bf4fafefe8cff88af88672f7110