Rename CJS_{KeyValue,GlobalData,GlobalVariableArray}.{h,cpp} to CFX_
Because these implement a typed value system akin to base::Value,
and do not call into any JavaScript at this layer. We can
reduce the number of simliar-sounding CJS_Global* names by using
the CFX_ prefix, and the CFX_ prefix hints that we should be
able someday to write a unit test for this without needing JS.

Rename CFX_GlobalVariableArray to CFX_GlobalArray for brevity.

Change-Id: I7e16760eb86c1690a68f8ac05c9a84a63f506556