Make CPDF_ContentMarkItem own a copy of its properties dict.
It could be aliased with some other dictionary in the file. We
note that the dictionary one level up will always be an indirect
object in the sharing case, and indirect objects are persisted
by the IndirectObjectHolder, so hold a pointer to that and retrieve
the specific property_name field on they fly.

Stick some |k| into constant names.  They look weird otherwise.

Bug: 900552
Change-Id: I2e300020d6a7191648dd139a485b6d284e259976