Make CPDF_ContentMarkItem stop caching the properties dict.
It could be aliased with some other dictionary in the file. We
note that the dictionary one level up will always be an indirect
object in the sharing case, and indirect objects are persisted
by the IndirectObjectHolder, so hold a pointer to that and retrieve
the specific property_name field on the fly.

Bug: chromium:900552
Change-Id: I2e300020d6a7191648dd139a485b6d284e259976