Move another chunk of code to fx_date_helper.cpp.
In this interim step, several methods have to become public as
they are used both locally and in cjs_publicmethods.cpp, as well
as the kMonth/kFullMonth arrays (now namespaced).

Remove the bWrongFormat out boolean param, and just use Nan
as a sentinel to indicate failure.

Move the non-JS parts of ParseDateUsingFormat as well.

Change-Id: Iff19a3aa94be39bf84d694f2b2e02224e0a36300