Speculative fix for unit test failure under code coverage
The last test that the bot attempts to execute before exiting
unexpectedly is CFDE_TextEditEngineTest.Insert. The set up
method for this test contains an ASSERT (ordinary assert, not
a special gtest assert) that a font was loaded. The guess is that
something about the environment on the coverage bot is causing
this operation to fail and the ASSERT to trip.

Change-Id: Ic215ece215c123d72ce162a3cf85a571bef8ace2