Update patch set 1
Patch Set 1:

We are writing past the end of a string, but not linearly. e.g. For "foo" and a block count of 2, it pads the string out to "foo____" and then fills it in in 2 passes: "fooN_N_", "fooNMNM".

Patch-set: 1
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false