Move packet info from cxfa_document_parser.cpp to xfa_basic_data.cpp.
Put it in the same file as the other generated table.

Introduce XFA_PACKETINFO struct to separate the results of a query
from the underlying table itself. Omit the hash from this struct
since it is an implementation detail that may someday go away. This
requires us to rehash a few strings, but the code that it calls with
this hash should someday get refactored to take a saner argument.

When looking up packets by name, do a final string comparison instead
of a hash comparison to avoid aliasing alternate forms of user input
to these names.

Add unit tests for API.

Change-Id: I7659ba79058591e1c0a4417c318cdcb696ba37c5