Update patch set 1
Patch Set 1:

Context for this CL: We enabled Tricium (Tricorder for Chroium, http://bit.ly/tricium) to run analyses and add robot comments to CLs in pdfium in https://pdfium-review.googlesource.com/c/pdfium/+/32731.

The purpose of this CL is to explicitly say that we don't want Tricium to analyze generated test expectation files here. (This is currently done via regexp in Tricium itself, but this specific rule is pdfium-specific so it's good to keep it here -- later if we want to disable Tricium for any other files, then it will be a small change in the relevant repo/directory).

Patch-set: 1
Reviewer: Gerrit User 5490 <5490@fca2d64a-0420-3532-b200-12848c95a47b>
CC: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 7245 <7245@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false