Check inputs for CPDFSDK_CustomAccess::ReadBlockAtOffset().
To be consistent with other IFX_SeekableReadStream::ReadBlockAtOffset()
implementation. Do the same checks in similar code that reads from
FPDF_FILEACCESS.

With these checks, CPDF_Parser::ParseAndAppendCrossRefSubsectionData()
will consistently fail for xref tables with 0 entries. Check for this
case and make the parser succeed. This deals with the xref table quirk
more gracefully and avoids rebuilds. Update test expectations as a
result of this new behavior.

BUG=pdfium:1197

Change-Id: Id9900161c3453c37b51d501cb8746618388214a9