Update patch set 2
Patch Set 2:

A better way to fix this is to change the documentation to state the password parameter should be UTF-8 encoded. To make the code backward compatible, for revisions that need Latin-1, test the password for the possibility it is already Latin-1 encoded. I.e. for revisions 5 and 6 handle it as now, for revision 4 and bellow try to decode and encode from UTF-8 to Latin-1, and test again, assuming already Latin-1, if the UTF-8 decoding fails or the re-encoded password is not valid.
How the password should be encoded is something for the security handler. Code external to the library should not need to know anything about this.

Patch-set: 2
CC: Gerrit User 7407 <7407@fca2d64a-0420-3532-b200-12848c95a47b>