Dis-entangle cfx_char.h from fx_bidi.h.
The one function in fx_bidi.h that references cfx_chars becomes a
static method in cfx_char.h. Then all of the other XFA-only code
moves to CFX_Char, which is appropriate since it is twiddling
vectors of these things anyways. Some ifdef XFAs go away in the
process.

Change-Id: I70c21be4865ceaa35852df77017f47cf1c58f440