Update patch set 1
Patch Set 1:

I'm going to add a test for this bug: https://pdfium-review.googlesource.com/48390

Right now, the test expectations are for the status quo, which is the incorrect behavior. Once the test has been committed, this CL will cause the test to fail. So this CL will need to be updated to set the correct expectations in the test.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>