Add CFXJSE_Value::IsEmpty() and use it in SetArray()
Because values have yet another possibility: they might be empty
and v8 will assert if we try to set an array element to an empty
value. Note that this is an entirely different state than being
undefined, or null, which it handles properly. In fact, newly
created CFXJSE_Values are IsEmpty().

Needed to fix a segv in a forthcoming CL.

Change-Id: I74f075ab3e483533e9c6ad81254e2b819553cdf8