Update patch set 2
Patch Set 2:

I generated a PDF file using PDFium. In that file, I had a TrueType font with a name like "Name With Spaces" embedded in.

The PDF file is fine to view with any PDF reader，but when I use Fontforge to extract the fonts in it, I see the font name is "Name#20With#20Spaces", which is encoded in the name encoding rule and Fontforge can't handle.

Then I refer to a PDF generated by MS Word and find the spaces in the font name are removed in that file.

Patch-set: 2
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 7520 <7520@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false