Update patch set 1
Patch Set 1:

It looks like you need the updates to build\vs_toolchain.py from crrev.com/c/1444031. That is, you have the updates to win_toolchain\get_toolchain_if_necessary.py in depot_tools so that it will only download a toolchain when explicitly requested, but you don't have the updates to build\vs_toolchain.py that pass that flag.

It should be a simple merge. Let me know if you hit any issues or if this doesn't make sense.

Patch-set: 1
CC: Gerrit User 6195 <6195@fca2d64a-0420-3532-b200-12848c95a47b>