Update patch set 1
Patch Set 1:

> What's weird here is the error message says to run gclient runhooks... inside the gclient runhooks step.

Gotcha. Okay, here's what must be happening.

pdfium/tools/clang/scripts/update.py is the missing link. It is asking for the toolchain without passing the 'update' parameter, so the toolchain script hasn't not been "authorized" to download the toolchain. So, either update.py needs to pass 'update' (assuming it is only called in the right context) or (since it requires that the toolchain be updated) it should be invoked after the toolchain update.

I'm not 100% clear on how 'gclient runhooks' works - is there a win_toolchain step which needs to be earlier. Sorry for the complication.

Patch-set: 1