Update patch set 1
Patch Set 1:

From the chromium DEPS file:

  {
    # Update the Windows toolchain if necessary.  Must run before 'clang' below.
    'name': 'win_toolchain',
    'pattern': '.',
    'condition': 'checkout_win',
    'action': ['python', 'src/build/vs_toolchain.py', 'update', '--force'],
  },
...
  {
    # Note: On Win, this should run after win_toolchain, as it may use it.
    'name': 'clang',
    'pattern': '.',
    'action': ['python', 'src/tools/clang/scripts/update.py', '--with-android={checkout_android}'],
  },

It's so important the win_toolchain run before clang that we commented it twice!

In the pdfium DEPS file the ordering is reversed. This is wrong, and has been wrong for a long time. I'm actually not surprised because the bugs that I fixed should never have happened on the bots unless they were misconfigured so I assumed that some misconfiguring would eventually be found.

So, it's good that the incorrect ordering has been found. Fixing it is probably a delicate operations, but important since otherwise we have unexpected behavior.

Try moving win_toolchain (and mac_toolchain?) in front of clang?

Patch-set: 1