Add null test before inserting in DoLayoutFlowedContainer()
The change at revision 3c90892 disallowed inserting null objects
into these linked lists which were previously silently tolerated,
so an explicit null check is required.

I suspect it was probably always wrong to take this branch with
a null object, but it was not noticed before due to the overly
tolerant list-insert code. In other words, we'd try to insert
after NULL, do nothing, presumably lose track of an object,
and leak.

Bug: pdfium:1234
Change-Id: Iaa12a44be98b246da26a3399bf0b34beaf2da74b