Add null test before inserting in DoLayoutFlowedContainer()
The change at revision 3c90892 disallowed inserting null objects
into these linked lists which were previously silently tolerated,
so an explicit null check is required.

I suspect it was probably always wrong to take this branch with
a null object, but it was not noticed before due to the ad-hoc
list-insert code. In other words, we'd terminate a list in its
middle, and even possibly lose track of any objects following it.

Bug: pdfium:1234
Change-Id: Iaa12a44be98b246da26a3399bf0b34beaf2da74b